Treat habitats as under-equipped EL workshops
Extraplanetary Launchpads marks any workshop with a productivity factor >= 1 as "fully equipped", expanding the set of kerbals who can work there and using different productivity formulas. The EL manual makes it clear that fully equipped status represents a space that's designed for construction work, and that generic crew spaces should not have it.