Add Soil RESOURCE to parts with soil recyclers
On the stock parts where the Snacks mod adds Soil recyclers, it also
adds Soil storage capacity, in a ratio of 50 Soil per crew capacity.
This makes sense since Soil is the recycler's input resource.

The greenhouses were already configured to hold 100 Soil per crew
capacity.  That hasn't changed, but it's now expressed in a simpler way
that makes it clear it's twice as much as other parts with recyclers
would have.